case-insenstive check on windows and undo info logging
